+++ index.html
@@ -445,6 +445,6 @@
       </aside>
     </div>
 
-    <script src="js/app.js" defer></script>
+    <script type="module" src="js/app.js"></script>
   </body>
 </html>
