+++ test/updateDrugDefaults.test.js
@@ -1,47 +1,40 @@
-const fs = require('fs');
-const vm = require('vm');
-const assert = require('assert');
+import assert from 'assert';
 
 const elements = {};
-function createEl(){ return { value: '', style: {}, addEventListener: ()=>{} }; }
-function getEl(key){ if(!elements[key]) elements[key] = createEl(); return elements[key]; }
+function createEl() { return { value: '', style: {}, addEventListener: () => {} }; }
+function getEl(key) { if (!elements[key]) elements[key] = createEl(); return elements[key]; }
 
 const documentStub = {
   querySelector: (sel) => getEl(sel),
   querySelectorAll: () => [],
-  getElementById: (id) => getEl('#'+id),
+  getElementById: (id) => getEl('#' + id),
   addEventListener: () => {},
-  createElement: () => createEl()
+  createElement: () => createEl(),
 };
 
-const sandbox = {
-  document: documentStub,
-  alert: ()=>{},
-  confirm: ()=>true,
-  localStorage: { setItem: ()=>{}, getItem: ()=>null },
-  URL: { createObjectURL: ()=>'', revokeObjectURL: ()=>{} },
-  Blob: function(){},
-  FileReader: function(){ this.readAsText = ()=>{}; },
-  setInterval: ()=>{}
-};
+global.document = documentStub;
+global.alert = () => {};
+global.confirm = () => true;
+global.localStorage = { setItem: () => {}, getItem: () => null };
+global.URL = { createObjectURL: () => '', revokeObjectURL: () => {} };
+global.Blob = function () {};
+global.FileReader = function () { this.readAsText = () => {}; };
+global.setInterval = () => {};
 
-vm.createContext(sandbox);
-const code = fs.readFileSync('js/app.js', 'utf8');
-vm.runInContext(code, sandbox);
-vm.runInContext('this.inputs = inputs; this.updateDrugDefaults = updateDrugDefaults;', sandbox);
+const { inputs } = await import('../js/state.js');
+const { updateDrugDefaults } = await import('../js/drugs.js');
 
 // When inputs for defaults are empty, updateDrugDefaults should
 // populate drug concentration with hard-coded defaults (5 for TNK, 1 for tPA).
-sandbox.inputs.def_tnk.value = '';
-sandbox.inputs.def_tpa.value = '';
+inputs.def_tnk.value = '';
+inputs.def_tpa.value = '';
 
-sandbox.inputs.drugType.value = 'tnk';
-sandbox.updateDrugDefaults();
-assert.strictEqual(sandbox.inputs.drugConc.value, '5');
+inputs.drugType.value = 'tnk';
+updateDrugDefaults();
+assert.strictEqual(inputs.drugConc.value, '5');
 
-sandbox.inputs.drugType.value = 'tpa';
-sandbox.updateDrugDefaults();
-assert.strictEqual(sandbox.inputs.drugConc.value, '1');
+inputs.drugType.value = 'tpa';
+updateDrugDefaults();
+assert.strictEqual(inputs.drugConc.value, '1');
 
 console.log('updateDrugDefaults sets default concentrations correctly');
-
